got it to run to completion
It only sort of works though. When you run it in training mode, if you
put files in the /TestFiles directory, it will test those as soon as it
finishes training. It's not very accurate and hilariously enough, has a
very hard time classifying unlambda. But if you put a single word like
"java" or "#include" it guesses pretty accurately. I messed with the
classifier math to get it to work and may have messed it up a little.